Implement pingback system

+++ src/middlewares/cleanup.ts
@@ -1,20 +1,22 @@
 import { Protocol, Response } from "restana"
-import { ExtendedRequest } from "../types"
 import path from "path"
 import rimraf from "rimraf"
+
+import { ExtendedRequest } from "../types"
+import queue from "../utils/queue"
 
 export default (
   req: ExtendedRequest,
   res: Response<Protocol.HTTPS>,
   next: () => void
 ): void => {
-  const dirToRemove = path.join(req.locals.cairoDir, "../..")
-  rimraf(dirToRemove, (err: Error): void => {
-    if (err) throw err
-    console.log(`success deleting ${dirToRemove}`)
-  })
+  if (!queue.length) {
+    const dirToRemove = path.join(req.locals.cairoDir, "../..")
+    rimraf(dirToRemove, (err: Error): void => {
+      if (err) throw err
+      console.log(`success deleting ${dirToRemove}`)
+    })
+  }
 
-  // Go to next middleware regardless so we
-  // can respond to client as soon as possible
   next()
 }
